Split out interface re-use for documentation
