+++ src/lib/dependencies.ts
@@ -423,8 +423,6 @@ function resolveTypeDependencyFrom (src: string, raw: string, options: Options, 
           main: config.main,
           version: config.version,
           browser: config.browser,
-          typings: config.typings,
-          browserTypings: config.browserTypings,
           files: Array.isArray(config.files) ? config.files : undefined,
           type: PROJECT_NAME,
           ambient: !!config.ambient,
